코드 추가

+++ app/src/main/java/org/sopt/sopkathon_31th/data/remote/api/profile/ProfileService.kt
@@ -8,6 +8,6 @@ import retrofit2.http.Path
 interface ProfileService {
     @GET("/page/{userId}")
     fun getProfile(
-        @Path("userId") userId: Long
+        @Path("userId") userId: Int
     ): Call<ResponseProfileDto>
 }
